# server/openapi-gamelibrary.yaml
openapi: 3.0.3
info:
  title: Game Library API - Group 34
  description: 'API Spec for the game library project for the course "Web Frameworks" for a back-end in Laravel.'
  version: 0.0.1
servers:
  - url: http://localhost:8000/
    description: Local API
...
paths:
...
  /games/{id}:
    get:
      summary: Returns a game from your library
      description: Returns a game from your library
      operationId: getGame
      tags:
        - Games
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the game to return
      responses:
        "200":
          description: "Got game"
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Game"
        "404":
          description: "Game not found"
        "500":
          description: "Unexpected error"
    put:
      summary: Update a game from your library
      description: Update a game from your library
      operationId: updateGame
      tags:
        - Games
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the game to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Game Title"
                price:
                  type: integer
                  example: 10
                description:
                  type: string
                  example: "Game description"
                rating:
                  type: integer
                  example: 8
                tags:
                  type: array
                  items:
                    type: string
                    example: "Action"
      responses:
        "200":
          description: "Game updated"
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Game"
        "400":
          description: "Bad request"
        "404":
          description: "Game not found"
        "500":
          description: "Unexpected error"
    delete:
      summary: Delete a game from your library
      description: Delete a game from your library
      operationId: deleteGame
      tags:
        - Games
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the game to delete
      responses:
        "204":
          description: "Game deleted"
        "404":
          description: "Game not found"
        "500":
          description: "Unexpected error"
...
components:
  schemas:
    Game:
      type: object
      properties:
        id:
          type: integer
          description: ID of the game
        title:
          type: string
          description: Title of the game
        rating:
          type: integer
          description: Rating of the game
        price:
          type: number
          description: Price of the game
        languages:
          type: array
          items:
            $ref: "#/components/schemas/GameLanguage"
    GameLanguage:
      type: object
      properties:
        language_id:
          type: integer
          description: ID of the language
        game_id:
          type: integer
          description: ID of the game
        language:
          type: string
          description: Language code
        description:
          type: string
          description: Description of the game in the specified language
        tags:
          type: array
          items:
            type: string
          description: Tags associated with the game in the specified language
...